Update index.html
Updated links for METARs, TAFs, BASH, and FWB. Aviation weather went through an update and now uses "," as separators for ICAO, instead of "+". FWB has a new link that no longer includes "/18" and BASH no longer resolves to a separate website when preferences are selected so I just changed the link to the BASH homepage.